Exercise 1.2.
While doing exercise 1.2, I noticed that there is a mistake in my
solution of exercise 1.1: the substitution rule Subst_2 actually isn't
what we're using; instead, we're using the *congruence* property of
judgemental equality, which isn't actually explicitly mentioned in the
appendix. More precisely, it is only mentioned in a throw-away remark,
and only for *constructors*, not eliminators (of which function
application is a special case).