Added `Issues.md`.
It is not entirely clear whether the rules in appendix A.2 are complete.
More precisely, it's not clear whether the congruence property is
assumed or follows given what's stated there.

Since I don't want to waste time, I simply assume this now in the
solution of exercise 1.1 and have added a new file (`Issues.md`) meant
to collect these various open questions/issues that will inevitably
arise, and with which I can't deal right now. Hopefully, at some point
in the future though.