Unforeseen complications.
Trying to make precise how the "formal calculation" in the solution of
exercise 1.2 stands for some derivation in Martin-Löf type theory has
turned out to be way more complicated than expected.
For example, only now I became aware of the fact that in a context, only
assumptions of the form $x : A$ for $x$ a variable (in the object
language) are allowed, which means that you cannot assume $a : A$ for
a *term* $a$ in the object language.