Convention for implicit arguments.
What makes formalization *seem* intractable is that the redundant
information---that which is suppressed in "normal", human mathematical
notation---is written out explicitly in the most naive approach.
Put differently, the process of elaboration, or rather its convers, the
suppression of redundant/implicit arguments is as necessary in formal
mathematics as it is in ordinary mathematics.