Finished exercise 1.6.
What looked like an easy (and boring) exercise turned out to be way more
involved than expected. That's due to the fact that carrying around
proof objects is somewhat cumbersome, especially when one tries to do
that "informally". In that sense, I am somewhat skeptical of the book's
premise of doing homotopy type theory in an informal way. To me it seems
like that this actually makes things harder than if you were to do this
formally in Agda.