Based path-induction via homotopy pullbacks.
My previous "geometrical-intuitive" sketch of a proof of how to deduce
based from unbased path-induction was nonsensical.

I realized that the correct way to proceed is to prove that the
canonical square diagram relating the free and based path-space is
a *homotopy* pullback square. To this end, I've introduced a definition
that shall be used in order to prove the necessary auxillary results in
their appropriate generality.