Homotopy pullback squares.
I've added a short discussion of homotopy pullback squares in
pre-categories and proved some simple lemmas. It remains to be shown
that the diagram between the free and based path spaces forms a homotopy
pullback square.

Unfortunately, I've noticed that the straightforward application of the homotopy pullback property only gives a weaker result than the desired one.